[CP-3194] 'ConnectFailed' event logic moved to main device-manager domain (mds logic)
