+++ libs/device-manager/feature/src/helpers/index.ts
@@ -3,4 +3,5 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
+export * from "./group-events-by-device-type"
 export * from "./is-core-device"
