+++ libs/device-manager/feature/src/hooks/index.ts
@@ -5,3 +5,4 @@
 
 export * from "./use-device-manager-connect-failed"
 export * from "./use-device-manager-connected"
+export * from "./use-device-manager-detached"
