+++ libs/generic-view/store/src/lib/hooks/index.ts
@@ -3,7 +3,6 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./use-api-serial-port-listeners"
 export * from "./use-outbox"
 export * from "./use-screen-title"
 export * from "./use-locked-device-handler"
