+++ libs/generic-view/ui/src/index.ts
@@ -27,7 +27,6 @@ export * from "./lib/predefined/backup-restore/backup-restore-error"
 export * from "./lib/predefined/import-contacts/import-contacts-error"
 export * from "./lib/predefined/data-migration/components/transfer-error-modal"
 export { DataMigrationPage } from "./lib/predefined/data-migration/data-migration"
-export { RecoveryModePage } from "msc-flash-harmony"
 export * from "./lib/buttons/button-text"
 export * from "./lib/buttons/button-primary"
 export * from "./lib/texts/paragraphs"
